Fix for self-destruction and waffling problems: Previously the automa considered all branches with assured failure to be of equal value, including branches where it ended the game itself eliminating its own pieces.
Similarly, it considered all branches with assured victory to be equal, which lead it to make 'waffling' moves where it would stretch out the game even though victory was assured.

Neither of these is strictly forbidden, but as a player they're both bad form and not fun to play against.

Previously attempted to check each move for validity, but a broader and simpler solution is to place more value on earlier victories. This is done by modifying a player victory score by the depth it is discovered at, thus making victories at different depth have different value, but still be better than any other possible board configuration.